Enhance Vulkan renderer and test infrastructure
- Added new Vulkan-related source files and resource management files to `CMakeLists.txt` targets.
- Integrated ImGui sources for `test_vulkan` and suppressed warnings for MSVC.
- Introduced new shader compilation steps for light culling, shadow rendering, and depth prepass.
- Refactored descriptor tests to support bindless textures and added tests for texture registration and unregistration.
- Updated `VulkanRenderer` initialization to include Forward+ light culling and shadow systems.
- Improved test executables with additional Vulkan files, updated dependencies, and renamed tests for clarity.
- Changed vertex structure to include alpha channel in `test_vulkan_buffer.cpp`.
- Simplified enum assertions in `test_mipmap_policy.cpp` and added type safety in `test_mipmap_array.cpp`.
- General cleanup and modernization for improved readability and maintainability.

+++ tests/test_mipmap_array.cpp
@@ -71,7 +71,7 @@ TextureData CreateTestArrayTexture(u32 width, u32 height, u32 channels, u32 laye
             for (u32 x = 0; x < width; ++x) {
                 u32 idx = (y * width + x) * channels;
                 bool checker = ((x / 16) + (y / 16)) % 2;
-                u8 value = checker ? (255 - i * 30) : (i * 30);
+                u8 value = static_cast<u8>(checker ? (255 - i * 30) : (i * 30));
 
                 for (u32 c = 0; c < channels; ++c) {
                     layerData[idx + c] = value;
